@@ -3,7 +3,7 @@ package br.com.fclug.financialaid.dialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.ColorInt;
+import androidx.annotation.ColorInt;
 import android.view.View;
 import android.view.Window;
 import android.widget.Button;
